Implements chifra slurp --type [...|withdrawals|...]

+++ src/apps/chifra/pkg/monitor/filter.go
@@ -63,6 +63,7 @@ func (mon *Monitor) ReadAndFilterAppearances(filt *filter.AppearanceFilter) (app
 	for _, app := range fromDisc {
 		app := app
 		if app.TransactionId == uint32(types.Withdrawal) {
+			// #WITHDRAWALS--THIS IS INCORRECT
 			continue
 		}
 
@@ -75,8 +76,8 @@ func (mon *Monitor) ReadAndFilterAppearances(filt *filter.AppearanceFilter) (app
 			filt.OuterBounds.Last = uint64(app.BlockNumber + 1)
 			s := types.SimpleAppearance{
 				Address:          mon.Address,
-				BlockNumber:      uint32(app.BlockNumber),
-				TransactionIndex: uint32(app.TransactionId),
+				BlockNumber:      app.BlockNumber,
+				TransactionIndex: app.TransactionId,
 				Timestamp:        utils.NOPOSI,
 			}
 			apps = append(apps, s)
